Make PostgreSQL data reproducible during E2E tests (#64)
Closes #63 

This MR makes all E2E tests start with the same data in the database,
making tests that modify the database possible. Here is a list of
changes:
- test parallelism is disabled, by limiting Playwright to 1 worker
process in the config
- `db/20-dump.sh` was added, which dumps the database data after
inserting the mock data
- `global.setup.ts` and `global.teardown.ts` were added, which save the
database state before any tests run and restore it once they all finish.
This ensures developers keep their current data they might have while
working on the app.
- The database is reset to the mock data dump before each test (page
fixture in `fixutres.ts`)

To dump the database data,
[pg_dump](https://www.postgresql.org/docs/17/app-pgdump.html) is used
which outputs an SQL file, that can then be run with `psql`.

---------

Co-authored-by: Adam Bucior <35536269+AdamBucior@users.noreply.github.com>