Add hot reload support (#61)
This PR makes hot reload possible for instantly previewing changes made
to the website.
Instructions for running the new development setup are located in the
README.
The way this was achieved is by replacing `express-session` with the new
[Session API](https://docs.astro.build/en/guides/sessions/) from
Astro.js. This enables us to use the Astro development server as it
doesn't require Express to be running.

---------

Co-authored-by: Adam Bucior <35536269+AdamBucior@users.noreply.github.com>