Laboratory class editing (#101)
This adds the ability to click on a laboratory class in the subject
calendar to view it's details. Additionally logged in users can edit the
class from this modal.

This is missing any sort of collision detection with other classes and
holidays, which is also largely missing from the generation of classes,
so it will likely be introduced in a separate MR for both of these.

---------

Co-authored-by: Adam Bucior <35536269+AdamBucior@users.noreply.github.com>