Show user email as an image (#108)
This PR adds the [canvas](https://www.npmjs.com/package/canvas) package
for drawing the user email to an image. There is an endpoint that sends
the image given a user id.

The package used has some native dependencies that make using it
slightly more complicated. There are prebuilt binaries for Linux
distributions using `glibc`, however the Docker image we currently use
is based on Alpine Linux, which uses `musl` instead. We could switch to
the Debian-based `node:22` image, where the build is pretty fast, but
the image size goes up to around `1.8 GB`. I modified our Dockerfile to
instead install the needed native dependencies and build the package
from source. This makes the build time slightly longer (around 1 min on
my machine), but the image size is around `650 MB`, so about `77%` less.
I'm not sure which way we want to go.

There are also some JS-only packages, like
[pureimage](https://www.npmjs.com/package/pureimage), though they are
slower and less maintained.

---------

Co-authored-by: Adam Bucior <35536269+AdamBucior@users.noreply.github.com>